hybrid/person/requirements.txt
aniso8601
click
colorama
dlib
dnspython
face-recognition
face-recognition-models
Flask
Flask-Cors
Flask-RESTful
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
opencv-python
Pillow
pymongo
python-dotenv
pytz
six
Werkzeug
